backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
supabase==2.10.0
pytrends==4.9.2
newsapi-python==0.2.7
google-search-results==2.4.2
httpx==0.28.0
python-dotenv==1.0.1
pydantic==2.10.3
ruff==0.8.0
mangum==0.19.0
